Bloom/bloom-backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
httpx
pydantic
python-multipart
